requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-groq
langchain_huggingface
pypdf
tiktoken
aiofiles
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
faiss-cpu
python-dotenv
sentence-transformers
flask

-e .